# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Digital Workplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: department
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Departments ────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema: { type: integer }
        - name: assigneeId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Documents ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Users ─────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, clerkId, email, firstName, lastName, role, status, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, hr_manager, team_leader, employee, guest]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blocked]
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
    UserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UserInput:
      type: object
      required: [email, firstName, lastName, role]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, hr_manager, team_leader, employee, guest]
        departmentId:
          type: ["integer", "null"]
        jobTitle:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, admin, hr_manager, team_leader, employee, guest]
        departmentId:
          type: ["integer", "null"]
        jobTitle:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blocked]

    # ── Departments ────────────────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, name, status, priority, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        dueDate:
          type: ["string", "null"]
          format: date
        progress:
          type: integer
          minimum: 0
          maximum: 100
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        departmentId:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        dueDate:
          type: ["string", "null"]
          format: date
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        departmentId:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        dueDate:
          type: ["string", "null"]
          format: date
        progress:
          type: integer
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        assigneeAvatar:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        estimatedHours:
          type: ["number", "null"]
        loggedHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        projectId:
          type: ["integer", "null"]
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        estimatedHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        estimatedHours:
          type: ["number", "null"]
        loggedHours:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string

    # ── Documents ─────────────────────────────────────────────────────────────
...