# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Workplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  clerkId: string;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  status: UserStatus;
  createdAt: Date;
  /** @nullable */
  lastLoginAt?: Date | null;
}